// services/task-service/types/task.d.ts
// deno-lint-ignore-file no-empty-interface
/* this code was generated by automated tool,
   should not edit by hand */

export interface TaskService {
    CreateTask(request: CreateTask): Promise<TaskResponse>;
    ReadTask(request: TaskId): Promise<TaskResponse>;
    UpdateTask(request: TaskItem): Promise<TaskResponse>;
    DeleteTask(request: TaskId): Promise<TaskResponse>;
    ListTasks(request: ListTasks): Promise<ListTasksResponse>;
}

export interface TaskItem {
    id?: string;
    title?: string;
    description?: string;
    completed?: boolean;
}

export interface TaskId {
    id?: string;
}

export interface CreateTask {
    title?: string;
    description?: string;
}

export interface TaskResponse {
    success?: boolean;
    task?: TaskItem;
}

export interface ListTasks {}

export interface ListTasksResponse {
    tasks?: TaskItem[];
}